+++ builder/src/main/java/com/stratio/cassandra/lucene/builder/Builder.java
@@ -760,8 +760,8 @@ public abstract class Builder {
     /**
      * Returns a new {@link Partitioner.None} to not partitioning the index.
      *
-     * Index partitioning is useful to speed up some queries to the detriment of others, depending on the implementation.
-     * It is also useful to overcome the Lucene's hard limit of 2147483519 documents per index.
+     * Index partitioning is useful to speed up some queries to the detriment of others, depending on the
+     * implementation. It is also useful to overcome the Lucene's hard limit of 2147483519 documents per index.
      *
      * @return a new no-action partitioning, equivalent to just don't partitioning the index
      */
@@ -795,10 +795,29 @@ public abstract class Builder {
      * cardinalities and uniform distributions will provide better load balancing between partitions.
      *
      * @param partitions the number of index partitions per node
-     * @param column     the name of the partition key column
+     * @param column the name of the partition key column
      * @return a new partitioner based on a partitioning key column
      */
     public static Partitioner.OnColumn partitionerOnColumn(int partitions, String column) {
         return new Partitioner.OnColumn(partitions, column);
     }
+
+    /**
+     * Returns a new {@link Partitioner.OnVirtualNode} based on the partition key token. Rows will be stored in an index
+     * partition determined by the virtual nodes token range. Partition-directed searches will be routed to a single
+     * partition, increasing performance. However, unbounded token range searches will be routed to all the partitions,
+     * with a slightly lower performance. Virtual node token range queries will be routed to only one partition which
+     * increase performance in spark queries with vnodes rather than partitioning on token.
+     *
+     * This partitioner load balance depends on virtual node token ranges assignation. The more virtual nodes, the
+     * better distribution (more similarity in number of tokens that falls inside any virtual node) between virtual
+     * nodes, the better load balance with this partitioner.
+     *
+     * @param virtualNodesPerPartition the number of virtual nodes per each partition
+     * @return a new partitioner based on Cassandra's vnode partitioning token ranges
+     */
+    public static Partitioner.OnVirtualNode partitionerOnVirtualNodes(int virtualNodesPerPartition) {
+        return new Partitioner.OnVirtualNode(virtualNodesPerPartition);
+    }
+
 }